+++ fastly/__init__.py
@@ -10,7 +10,7 @@
 """
 
 
-__version__ = "2.2.0"
+__version__ = "2.2.1"
 
 # import ApiClient
 from fastly.api_client import ApiClient