+++ fastly/model/relationship_tls_configuration_for_tls_subscription.py
@@ -30,25 +30,11 @@ from fastly.exceptions import ApiAttributeError
 
 
 def lazy_import():
-    from fastly.model.relationship_common_name import RelationshipCommonName
-    from fastly.model.relationship_member_tls_domain import RelationshipMemberTlsDomain
-    from fastly.model.relationship_tls_certificates import RelationshipTlsCertificates
-    from fastly.model.relationship_tls_certificates_tls_certificates import RelationshipTlsCertificatesTlsCertificates
-    from fastly.model.relationship_tls_configuration_for_tls_subscription import RelationshipTlsConfigurationForTlsSubscription
     from fastly.model.relationship_tls_configuration_tls_configuration import RelationshipTlsConfigurationTlsConfiguration
-    from fastly.model.relationship_tls_domains import RelationshipTlsDomains
-    from fastly.model.relationship_tls_domains_tls_domains import RelationshipTlsDomainsTlsDomains
-    globals()['RelationshipCommonName'] = RelationshipCommonName
-    globals()['RelationshipMemberTlsDomain'] = RelationshipMemberTlsDomain
-    globals()['RelationshipTlsCertificates'] = RelationshipTlsCertificates
-    globals()['RelationshipTlsCertificatesTlsCertificates'] = RelationshipTlsCertificatesTlsCertificates
-    globals()['RelationshipTlsConfigurationForTlsSubscription'] = RelationshipTlsConfigurationForTlsSubscription
     globals()['RelationshipTlsConfigurationTlsConfiguration'] = RelationshipTlsConfigurationTlsConfiguration
-    globals()['RelationshipTlsDomains'] = RelationshipTlsDomains
-    globals()['RelationshipTlsDomainsTlsDomains'] = RelationshipTlsDomainsTlsDomains
 
 
-class RelationshipsForTlsSubscription(ModelComposed):
+class RelationshipTlsConfigurationForTlsSubscription(ModelNormal):
     """NOTE: This class is auto generated.
     Do not edit the class manually.
 
@@ -100,9 +86,6 @@ class RelationshipsForTlsSubscription(ModelComposed):
         lazy_import()
         return {
             'tls_configuration': (RelationshipTlsConfigurationTlsConfiguration,),  # noqa: E501
-            'common_name': (RelationshipMemberTlsDomain,),  # noqa: E501
-            'tls_domains': (RelationshipTlsDomainsTlsDomains,),  # noqa: E501
-            'tls_certificates': (RelationshipTlsCertificatesTlsCertificates,),  # noqa: E501
         }
 
     @cached_property
@@ -112,18 +95,17 @@ class RelationshipsForTlsSubscription(ModelComposed):
 
     attribute_map = {
         'tls_configuration': 'tls_configuration',  # noqa: E501
-        'common_name': 'common_name',  # noqa: E501
-        'tls_domains': 'tls_domains',  # noqa: E501
-        'tls_certificates': 'tls_certificates',  # noqa: E501
     }
 
     read_only_vars = {
     }
 
+    _composed_schemas = {}
+
     @classmethod
     @convert_js_args_to_python_args
     def _from_openapi_data(cls, *args, **kwargs):  # noqa: E501
-        """RelationshipsForTlsSubscription - a model defined in OpenAPI
+        """RelationshipTlsConfigurationForTlsSubscription - a model defined in OpenAPI
 
         Keyword Args:
             _check_type (bool): if True, values for parameters in openapi_types
@@ -157,9 +139,6 @@ class RelationshipsForTlsSubscription(ModelComposed):
                                 through its discriminator because we passed in
                                 _visited_composed_classes = (Animal,)
             tls_configuration (RelationshipTlsConfigurationTlsConfiguration): [optional]  # noqa: E501
-            common_name (RelationshipMemberTlsDomain): [optional]  # noqa: E501
-            tls_domains (RelationshipTlsDomainsTlsDomains): [optional]  # noqa: E501
-            tls_certificates (RelationshipTlsCertificatesTlsCertificates): [optional]  # noqa: E501
         """
 
         _check_type = kwargs.pop('_check_type', True)
@@ -187,29 +166,14 @@ class RelationshipsForTlsSubscription(ModelComposed):
         self._configuration = _configuration
         self._visited_composed_classes = _visited_composed_classes + (self.__class__,)
 
-        constant_args = {
-            '_check_type': _check_type,
-            '_path_to_item': _path_to_item,
-            '_spec_property_naming': _spec_property_naming,
-            '_configuration': _configuration,
-            '_visited_composed_classes': self._visited_composed_classes,
-        }
-        composed_info = validate_get_composed_info(
-            constant_args, kwargs, self)
-        self._composed_instances = composed_info[0]
-        self._var_name_to_model_instances = composed_info[1]
-        self._additional_properties_model_instances = composed_info[2]
-        discarded_args = composed_info[3]
-
         for var_name, var_value in kwargs.items():
-            if var_name in discarded_args and \
+            if var_name not in self.attribute_map and \
                         self._configuration is not None and \
                         self._configuration.discard_unknown_keys and \
-                        self._additional_properties_model_instances:
+                        self.additional_properties_type is None:
                 # discard variable.
                 continue
             setattr(self, var_name, var_value)
-
         return self
 
     required_properties = set([
@@ -219,14 +183,11 @@ class RelationshipsForTlsSubscription(ModelComposed):
         '_path_to_item',
         '_configuration',
         '_visited_composed_classes',
-        '_composed_instances',
-        '_var_name_to_model_instances',
-        '_additional_properties_model_instances',
     ])
 
     @convert_js_args_to_python_args
     def __init__(self, *args, **kwargs):  # noqa: E501
-        """RelationshipsForTlsSubscription - a model defined in OpenAPI
+        """RelationshipTlsConfigurationForTlsSubscription - a model defined in OpenAPI
 
         Keyword Args:
             _check_type (bool): if True, values for parameters in openapi_types
@@ -260,9 +221,6 @@ class RelationshipsForTlsSubscription(ModelComposed):
                                 through its discriminator because we passed in
                                 _visited_composed_classes = (Animal,)
             tls_configuration (RelationshipTlsConfigurationTlsConfiguration): [optional]  # noqa: E501
-            common_name (RelationshipMemberTlsDomain): [optional]  # noqa: E501
-            tls_domains (RelationshipTlsDomainsTlsDomains): [optional]  # noqa: E501
-            tls_certificates (RelationshipTlsCertificatesTlsCertificates): [optional]  # noqa: E501
         """
 
         _check_type = kwargs.pop('_check_type', True)
@@ -288,51 +246,14 @@ class RelationshipsForTlsSubscription(ModelComposed):
         self._configuration = _configuration
         self._visited_composed_classes = _visited_composed_classes + (self.__class__,)
 
-        constant_args = {
-            '_check_type': _check_type,
-            '_path_to_item': _path_to_item,
-            '_spec_property_naming': _spec_property_naming,
-            '_configuration': _configuration,
-            '_visited_composed_classes': self._visited_composed_classes,
-        }
-        composed_info = validate_get_composed_info(
-            constant_args, kwargs, self)
-        self._composed_instances = composed_info[0]
-        self._var_name_to_model_instances = composed_info[1]
-        self._additional_properties_model_instances = composed_info[2]
-        discarded_args = composed_info[3]
-
         for var_name, var_value in kwargs.items():
-            if var_name in discarded_args and \
+            if var_name not in self.attribute_map and \
                         self._configuration is not None and \
                         self._configuration.discard_unknown_keys and \
-                        self._additional_properties_model_instances:
+                        self.additional_properties_type is None:
                 # discard variable.
                 continue
             setattr(self, var_name, var_value)
             if var_name in self.read_only_vars:
                 raise ApiAttributeError(f"`{var_name}` is a read-only attribute. Use `from_openapi_data` to instantiate "
                                      f"class with read only attributes.")
-
-    @cached_property
-    def _composed_schemas():
-        # we need this here to make our import statements work
-        # we must store _composed_schemas in here so the code is only run
-        # when we invoke this method. If we kept this at the class
-        # level we would get an error because the class level
-        # code would be run when this module is imported, and these composed
-        # classes don't exist yet because their module has not finished
-        # loading
-        lazy_import()
-        return {
-          'anyOf': [
-              RelationshipCommonName,
-              RelationshipTlsCertificates,
-              RelationshipTlsConfigurationForTlsSubscription,
-              RelationshipTlsDomains,
-          ],
-          'allOf': [
-          ],
-          'oneOf': [
-          ],
-        }